[#1742] Rename IntrospectionTranslator to PlainStructTranslator
