+++ integrations/ros2/workspace/src/demo_nodes/end-to-end-tests/prefix_mapping_plain_struct_translator_publisher.exp
@@ -20,8 +20,6 @@ source ${REPO_ROOT}/internal/end-to-end-testing/common.exp
 
 #### Test Setup
 
-set MAPPING integrations/ros2/workspace/src/demo_nodes/static_mapping_cmdvel.toml
-
 exec cargo build --manifest-path integrations/ros2/Cargo.toml \
     --bin iox2-tunnel-ros2 2>@1
 
@@ -29,18 +27,15 @@ exec cargo build --manifest-path integrations/ros2/Cargo.toml \
 
 set timeout 30
 
-# The mapping pairs the example's service `CmdVel` with the ROS 2 topic
-# `/cmd_vel`, kept separate from the chatter mapping because the introspection
-# translator admits fixed-size types only.
+# The example publishes on `ros2://topics/cmd_vel`, which the tunnel maps to
+# the ROS 2 topic `/cmd_vel` by the name prefix without further configuration.
 spawn cargo run --manifest-path integrations/ros2/Cargo.toml \
-    --bin iox2-tunnel-ros2 -- \
-    --static-mapping ${MAPPING} \
-    --translator introspection
+    --bin iox2-tunnel-ros2 -- --translator PlainStruct
 set id_tunnel $spawn_id
 expect_output_from $id_tunnel "Local Discovery via Tracker"
 
 spawn ros2 run demo_nodes_iceoryx2 \
-    static_mapping_introspection_translator_publisher
+    prefix_mapping_plain_struct_translator_publisher
 set id_publisher $spawn_id
 expect_output_from $id_publisher "send:"
