+++ integrations/ros2/workspace/src/demo_nodes/end-to-end-tests/static_mapping_plain_struct_translator_subscriber.exp
@@ -34,12 +34,12 @@ set timeout 30
 spawn cargo run --manifest-path integrations/ros2/Cargo.toml \
     --bin iox2-tunnel-ros2 -- \
     --static-mapping ${MAPPING} \
-    --translator introspection
+    --translator PlainStruct
 set id_tunnel $spawn_id
 expect_output_from $id_tunnel "Local Discovery via Tracker"
 
 spawn ros2 run demo_nodes_iceoryx2 \
-    static_mapping_introspection_translator_subscriber
+    static_mapping_plain_struct_translator_subscriber
 set id_subscriber $spawn_id
 expect_output_from $id_subscriber "waiting for messages on CmdVel"
