+++ integrations/ros2/workspace/src/demo_nodes/src/bin/static_mapping_plain_struct_translator_publisher.rs
@@ -12,15 +12,15 @@
 
 //! Publishes `geometry_msgs/msg/Twist` on the service statically mapped to
 //! the ROS 2 topic `/cmd_vel`. The application handles only the native
-//! [`Twist`] struct; the tunnel's introspection translator does the CDR
+//! [`Twist`] struct; the tunnel's plain-struct translator does the CDR
 //! (de)serialization.
 //!
 //! ```bash
-//! ros2 run demo_nodes_iceoryx2 static_mapping_introspection_translator_publisher
+//! ros2 run demo_nodes_iceoryx2 static_mapping_plain_struct_translator_publisher
 //! # in other shells:
 //! #   cargo run --bin iox2-tunnel-ros2 -- \
 //! #       --static-mapping workspace/src/demo_nodes/static_mapping_cmdvel.toml \
-//! #       --translator introspection
+//! #       --translator PlainStruct
 //! #   ros2 topic echo /cmd_vel
 //! ```
